商品削除機能　提出２回目

+++ app/controllers/products_controller.rb
@@ -24,8 +24,10 @@ class ProductsController < ApplicationController
 
   def destroy
     product = Product.find(params[:id])
-    product.destroy
-    redirect_to root_path
+    if user_signed_in? && current_user.id == product.user.id 
+      product.destroy
+      redirect_to root_path
+    end
   end
 
   private
